# Further OpenCL kernel improvements