This patch improves OpenCL kernel in the following aspects:
1. Shrink blake2b_state structure;
2. Eliminate the cost of table look-up;
3. Explict data access for efficiency;
4. Avoid unnecessary initializations;