@jserv 2020-08-02
@Inkeliz, can you please do some comparison between commit c7f05fa19b819c6c3b8c0bd6f55c4e1531db3a47 and this pull request with your setup?