@Inkeliz 2020-08-04
I'm using that code:

```
func BenchmarkGenerateWork(b *testing.B) {
	b.StopTimer()
	rand.Seed(1) // To make always test the same

	worker, err := NewWorkerGPU()
	if err != nil {
		panic(err)
	}
	
	DefaultWorkerPool = NewPool(worker) // To only use GPU (removing it will use CPU+GPU)

	for n := 0; n < b.N; n++ {
		b.StopTimer()
		hash, difficulty := make([]byte, 32), V2BaseDifficult
		rand.Read(hash)

		b.StartTimer()
		start := time.Now()
		_, err := GenerateWork(hash, difficulty)
		if err != nil {
			b.Error(err)
		}
		end := time.Now().Sub(start)
		b.StopTimer()

		fmt.Println(end)
	}

	b.StopTimer()
}
```

Just creating an `gpu_cl_test.go` and then execute:

> go test -tags cl -bench BenchmarkGenerateWork2 -benchtime 20x -timeout 1h

------

Since it needs multiples samples, I make that other one to store the results:

```
func BenchmarkGenerateWork(b *testing.B) {
	b.StopTimer()
	rand.Seed(1) // To make always test the same

	worker, err := NewWorkerGPU()
	if err != nil {
		b.Error(err)
	}

	DefaultWorkerPool = NewPool(worker)

	Start()

	for n := 0; n < b.N; n++ {
		b.StopTimer()
		hash, difficulty := make([]byte, 32), V2BaseDifficult
		rand.Read(hash)

		b.StartTimer()
		start := time.Now()
		_, err := GenerateWork(hash, difficulty)
		if err != nil {
			b.Error(err)
		}
		end := time.Now().Sub(start)
		b.StopTimer()

		Append(n, end)
	}

	b.StopTimer()

	End()

	return
}

var Results = make(map[int][]time.Duration, 1<<11)

func Start() {
	file, err := os.Open("gpu_cl_test_results.json")
	if err != nil {
		if os.IsNotExist(err) {
			return
		}
		panic(err)
	}

	defer file.Close()
	if err := json.NewDecoder(file).Decode(&Results); err != nil {
		panic(err)
	}
}

func Append(i int, d time.Duration) {
	if _, ok := Results[i]; !ok {
		Results[i] = make([]time.Duration, 0)
	}

	Results[i] = append(Results[i], d)
}

func End() {
	fileJSON, err := os.Create("gpu_cl_test_results.json")
	if err != nil {
		panic(err)
	}

	if err := json.NewEncoder(fileJSON).Encode(Results); err != nil {
		panic(err)
	}

	fileMD, err := os.Create("gpu_cl_test_results.md")
	if err != nil {
		panic(err)
	}

	if _, ok := Results[0]; !ok {
		return
	}

	for k := range Results[0] {
		fmt.Fprintf(fileMD, `| %d `, k)
	}
	fmt.Fprint(fileMD, "| ** AVG** | \n")

	for range Results[0] {
		fmt.Fprint(fileMD, `| -- `)
	}
	fmt.Fprint(fileMD, "| -- | \n")

	for i := 0; i < len(Results); i++ {
		total := int64(0)
		for _, vd := range Results[i] {
			fmt.Fprintf(fileMD, `| %s`, vd)
			total += vd.Nanoseconds()
		}

		for j := 0; j < (len(Results[0]) - len(Results[i])); j++ {
			fmt.Fprint(fileMD, `| - `)
		}

		fmt.Fprintf(fileMD, "| **%s** | \n", time.Duration(total / int64(len(Results[i]))))
	}
}
```

It's messy, but saves the result into `gpu_cl_test_results.json` and create the table at `gpu_cl_test_results.md`, which includes the AVG. The `time` is monotonic, so seems safe for such usage, I guess. I'll post the result using the new and the old one, I couldn't do that before because I can't use the device while computing the PoW. :\

But, thank you for the PR in advance. 👍