@Inkeliz 2020-08-04
WOW. I did some tests now, it's seems to be really faster now. I really need to take a look on that. Because it's far faster than the old one. It's so fast that I'm actually thinking that is something wrong, or it was a really nice job. 

It drops from 1.48719817 seconds per PoW to 0.299763834 seconds. Generating 100 PoWs takes 154.860s before, using only GPU, and now (with that new change) took 33.855s. I even put some `if !IsValid()`, all PoWs is generated are valid and uses the 8x difficulty.

---

I'll keep testing it futher.