@Inkeliz 2020-08-04
I test much time, and the results are almost always the same, even with other seeds (at `rand.Seed()`). All generated PoWs are valid against `IsValid` and also against `work_validate`. I'm posting here just some of the results. Otherwise, the table will be so big. The "AVG" uses all the data that I have collected, not only those 3 results.

----------

| Version | 1 | 2 | 3 | AVG | AVG DIFF | AVG DIFF % |
| -- | -- | -- | -- | -- | -- | -- | 
| Old | 0.862 | 0.862 | 0.863 | **0.861s**|  |  |
| New | 0.174 | 0.169 | 0.172 | **0.172s** | **-0.689s** | **-80.06%** |
| Old | 1.206 | 1.208 | 1.206 | **1.204s**|  |  |
| New | 0.240 | 0.244 | 0.240 | **0.242s** | **-0.962s** | **-79.88%** |
| Old | 0.830 | 0.832 | 0.830 | **0.829s**|  |  |
| New | 0.168 | 0.170 | 0.169 | **0.169s** | **-0.660s** | **-79.59%** |
| Old | 1.557 | 1.556 | 1.556 | **1.554s**|  |  |
| New | 0.316 | 0.315 | 0.311 | **0.314s** | **-1.240s** | **-79.82%** |
| Old | 0.988 | 0.979 | 0.980 | **0.980s**|  |  |
| New | 0.197 | 0.199 | 0.196 | **0.198s** | **-0.782s** | **-79.78%** |
| Old | 1.645 | 1.630 | 1.638 | **1.632s**|  |  |
| New | 0.325 | 0.327 | 0.326 | **0.327s** | **-1.305s** | **-79.96%** |
| Old | 3.100 | 3.057 | 3.059 | **3.063s**|  |  |
| New | 0.613 | 0.604 | 0.606 | **0.609s** | **-2.454s** | **-80.12%** |
| Old | 0.537 | 0.529 | 0.531 | **0.532s**|  |  |
| New | 0.110 | 0.109 | 0.109 | **0.109s** | **-0.422s** | **-79.41%** |
| Old | 3.214 | 3.210 | 3.207 | **3.208s**|  |  |
| New | 0.640 | 0.635 | 0.637 | **0.639s** | **-2.569s** | **-80.09%** |
| Old | 4.412 | 4.405 | 4.404 | **4.405s**|  |  |
| New | 0.871 | 0.869 | 0.880 | **0.876s** | **-3.529s** | **-80.11%** |
| Old | 1.854 | 1.855 | 1.856 | **1.855s**|  |  |
| New | 0.369 | 0.368 | 0.378 | **0.374s** | **-1.482s** | **-79.86%** |
| Old | 1.311 | 1.305 | 1.305 | **1.306s**|  |  |
| New | 0.260 | 0.260 | 0.279 | **0.266s** | **-1.041s** | **-79.66%** |
| Old | 9.644 | 9.629 | 9.630 | **9.631s**|  |  |
| New | 1.910 | 1.914 | 1.950 | **1.922s** | **-7.709s** | **-80.05%** |
| Old | 0.106 | 0.105 | 0.106 | **0.106s**|  |  |
| New | 0.024 | 0.024 | 0.024 | **0.025s** | **-0.081s** | **-76.57%** |
| Old | 1.556 | 1.555 | 1.557 | **1.555s**|  |  |
| New | 0.313 | 0.312 | 0.312 | **0.313s** | **-1.242s** | **-79.89%** |
| Old | 0.655 | 0.656 | 0.655 | **0.655s**|  |  |
| New | 0.135 | 0.133 | 0.134 | **0.134s** | **-0.521s** | **-79.52%** |
| Old | 0.157 | 0.155 | 0.156 | **0.156s**|  |  |
| New | 0.035 | 0.035 | 0.035 | **0.035s** | **-0.121s** | **-77.42%** |
| Old | 0.831 | 0.831 | 0.830 | **0.831s**|  |  |
| New | 0.170 | 0.167 | 0.175 | **0.170s** | **-0.661s** | **-79.55%** |
| Old | 3.299 | 3.279 | 3.282 | **3.283s**|  |  |
| New | 0.661 | 0.649 | 0.675 | **0.659s** | **-2.624s** | **-79.93%** |
| Old | 0.688 | 0.682 | 0.680 | **0.682s**|  |  |
| New | 0.140 | 0.138 | 0.142 | **0.140s** | **-0.542s** | **-79.47%** |
| Old | 3.779 | 3.756 | 3.757 | **3.760s**|  |  |
| New | 0.745 | 0.743 | 0.767 | **0.752s** | **-3.008s** | **-80%** |
| Old | 1.260 | 1.256 | 1.256 | **1.256s**|  |  |
| New | 0.251 | 0.250 | 0.257 | **0.253s** | **-1.003s** | **-79.84%** |
| Old | 1.444 | 1.430 | 1.430 | **1.433s**|  |  |
| New | 0.287 | 0.285 | 0.295 | **0.290s** | **-1.143s** | **-79.79%** |
| Old | 0.435 | 0.430 | 0.430 | **0.431s**|  |  |
| New | 0.092 | 0.090 | 0.090 | **0.090s** | **-0.341s** | **-79.06%** |
| Old | 1.655 | 1.630 | 1.631 | **1.634s**|  |  |
| New | 0.332 | 0.325 | 0.340 | **0.331s** | **-1.303s** | **-79.75%** |
| Old | 1.542 | 1.531 | 1.531 | **1.532s**|  |  |
| New | 0.320 | 0.305 | 0.315 | **0.312s** | **-1.221s** | **-79.67%** |
| Old | 0.606 | 0.606 | 0.607 | **0.606s**|  |  |
| New | 0.126 | 0.123 | 0.125 | **0.125s** | **-0.481s** | **-79.39%** |
| Old | 1.205 | 1.206 | 1.205 | **1.205s**|  |  |
| New | 0.245 | 0.241 | 0.244 | **0.244s** | **-0.961s** | **-79.76%** |
| Old | 2.339 | 2.330 | 2.333 | **2.330s**|  |  |
| New | 0.462 | 0.466 | 0.471 | **0.467s** | **-1.863s** | **-79.94%** |
| Old | 1.848 | 1.831 | 1.831 | **1.832s**|  |  |
| New | 0.365 | 0.369 | 0.368 | **0.368s** | **-1.464s** | **-79.93%** |
| Old | 0.867 | 0.856 | 0.856 | **0.857s**|  |  |
| New | 0.176 | 0.175 | 0.176 | **0.175s** | **-0.682s** | **-79.59%** |
| Old | 0.984 | 0.955 | 0.955 | **0.959s**|  |  |
| New | 0.192 | 0.195 | 0.196 | **0.194s** | **-0.765s** | **-79.74%** |
| Old | 1.940 | 1.929 | 1.931 | **1.930s**|  |  |
| New | 0.387 | 0.386 | 0.398 | **0.389s** | **-1.541s** | **-79.85%** |
| Old | 0.406 | 0.406 | 0.406 | **0.406s**|  |  |
| New | 0.086 | 0.085 | 0.084 | **0.085s** | **-0.321s** | **-79.12%** |
| Old | 0.130 | 0.131 | 0.130 | **0.130s**|  |  |
| New | 0.030 | 0.029 | 0.030 | **0.030s** | **-0.100s** | **-76.93%** |
| Old | 0.386 | 0.380 | 0.380 | **0.381s**|  |  |
| New | 0.080 | 0.079 | 0.081 | **0.080s** | **-0.302s** | **-79.05%** |
| Old | 0.081 | 0.080 | 0.081 | **0.081s**|  |  |
| New | 0.020 | 0.020 | 0.021 | **0.020s** | **-0.061s** | **-75.57%** |
| Old | 0.835 | 0.830 | 0.830 | **0.831s**|  |  |
| New | 0.170 | 0.167 | 0.170 | **0.169s** | **-0.662s** | **-79.66%** |
| Old | 1.969 | 1.955 | 1.957 | **1.960s**|  |  |
| New | 0.397 | 0.390 | 0.396 | **0.394s** | **-1.567s** | **-79.92%** |
| Old | 2.620 | 2.606 | 2.606 | **2.608s**|  |  |
| New | 0.519 | 0.518 | 0.521 | **0.520s** | **-2.088s** | **-80.05%** |
| Old | 0.130 | 0.130 | 0.131 | **0.130s**|  |  |
| New | 0.030 | 0.029 | 0.030 | **0.030s** | **-0.100s** | **-76.8%** |
| Old | 0.511 | 0.505 | 0.505 | **0.505s**|  |  |
| New | 0.105 | 0.104 | 0.105 | **0.104s** | **-0.401s** | **-79.34%** |
| Old | 2.415 | 2.406 | 2.406 | **2.406s**|  |  |
| New | 0.478 | 0.478 | 0.496 | **0.484s** | **-1.921s** | **-79.86%** |
| Old | 1.231 | 1.232 | 1.230 | **1.231s**|  |  |
| New | 0.246 | 0.247 | 0.255 | **0.250s** | **-0.982s** | **-79.73%** |
| Old | 0.055 | 0.055 | 0.056 | **0.055s**|  |  |
| New | 0.016 | 0.016 | 0.016 | **0.016s** | **-0.039s** | **-71.47%** |
| Old | 1.132 | 1.130 | 1.131 | **1.131s**|  |  |
| New | 0.226 | 0.226 | 0.235 | **0.229s** | **-0.901s** | **-79.7%** |
| Old | 0.358 | 0.355 | 0.356 | **0.356s**|  |  |
| New | 0.074 | 0.074 | 0.076 | **0.075s** | **-0.281s** | **-78.92%** |
| Old | 0.735 | 0.730 | 0.730 | **0.730s**|  |  |
| New | 0.148 | 0.148 | 0.151 | **0.149s** | **-0.581s** | **-79.55%** |
| Old | 1.156 | 1.155 | 1.157 | **1.155s**|  |  |
| New | 0.231 | 0.232 | 0.243 | **0.235s** | **-0.920s** | **-79.62%** |
| Old | 3.289 | 3.281 | 3.280 | **3.281s**|  |  |
| New | 0.651 | 0.650 | 0.659 | **0.656s** | **-2.625s** | **-80%** |
| Old | 0.607 | 0.605 | 0.606 | **0.605s**|  |  |
| New | 0.124 | 0.124 | 0.127 | **0.125s** | **-0.480s** | **-79.36%** |
| Old | 1.282 | 1.279 | 1.282 | **1.280s**|  |  |
| New | 0.257 | 0.260 | 0.259 | **0.259s** | **-1.021s** | **-79.77%** |
| Old | 0.555 | 0.555 | 0.555 | **0.555s**|  |  |
| New | 0.114 | 0.115 | 0.115 | **0.115s** | **-0.440s** | **-79.28%** |
| Old | 0.630 | 0.632 | 0.631 | **0.630s**|  |  |
| New | 0.128 | 0.130 | 0.131 | **0.130s** | **-0.500s** | **-79.4%** |
| Old | 2.738 | 2.731 | 2.731 | **2.732s**|  |  |
| New | 0.543 | 0.550 | 0.546 | **0.547s** | **-2.185s** | **-79.98%** |
| Old | 0.955 | 0.955 | 0.957 | **0.956s**|  |  |
| New | 0.193 | 0.192 | 0.193 | **0.194s** | **-0.762s** | **-79.74%** |
| Old | 0.355 | 0.355 | 0.354 | **0.355s**|  |  |
| New | 0.074 | 0.074 | 0.075 | **0.075s** | **-0.280s** | **-78.91%** |
| Old | 1.731 | 1.731 | 1.731 | **1.730s**|  |  |
| New | 0.350 | 0.344 | 0.346 | **0.348s** | **-1.383s** | **-79.91%** |
| Old | 1.562 | 1.554 | 1.555 | **1.555s**|  |  |
| New | 0.316 | 0.312 | 0.327 | **0.316s** | **-1.239s** | **-79.65%** |
| Old | 2.357 | 2.357 | 2.358 | **2.356s**|  |  |
| New | 0.476 | 0.468 | 0.480 | **0.474s** | **-1.882s** | **-79.88%** |
| Old | 1.130 | 1.130 | 1.130 | **1.131s**|  |  |
| New | 0.232 | 0.226 | 0.243 | **0.232s** | **-0.899s** | **-79.47%** |
| Old | 1.136 | 1.129 | 1.130 | **1.134s**|  |  |
| New | 0.228 | 0.227 | 0.231 | **0.230s** | **-0.905s** | **-79.76%** |
| Old | 0.282 | 0.280 | 0.281 | **0.282s**|  |  |
| New | 0.060 | 0.060 | 0.060 | **0.060s** | **-0.222s** | **-78.72%** |
| Old | 4.835 | 4.832 | 4.831 | **4.831s**|  |  |
| New | 0.957 | 0.955 | 0.974 | **0.968s** | **-3.863s** | **-79.96%** |
| Old | 0.030 | 0.029 | 0.030 | **0.030s**|  |  |
| New | 0.010 | 0.011 | 0.010 | **0.010s** | **-0.020s** | **-66.21%** |
| Old | 1.682 | 1.680 | 1.683 | **1.684s**|  |  |
| New | 0.336 | 0.334 | 0.336 | **0.338s** | **-1.346s** | **-79.95%** |
| Old | 0.380 | 0.380 | 0.379 | **0.382s**|  |  |
| New | 0.078 | 0.079 | 0.079 | **0.080s** | **-0.302s** | **-79.1%** |
| Old | 3.413 | 3.405 | 3.406 | **3.412s**|  |  |
| New | 0.676 | 0.676 | 0.687 | **0.683s** | **-2.729s** | **-79.99%** |
| Old | 2.257 | 2.258 | 2.256 | **2.258s**|  |  |
| New | 0.449 | 0.455 | 0.460 | **0.455s** | **-1.802s** | **-79.83%** |
| Old | 0.905 | 0.906 | 0.906 | **0.907s**|  |  |
| New | 0.182 | 0.185 | 0.185 | **0.184s** | **-0.723s** | **-79.67%** |
| Old | 0.280 | 0.280 | 0.281 | **0.281s**|  |  |
| New | 0.060 | 0.060 | 0.061 | **0.060s** | **-0.221s** | **-78.65%** |
| Old | 0.582 | 0.581 | 0.580 | **0.580s**|  |  |
| New | 0.118 | 0.120 | 0.122 | **0.120s** | **-0.460s** | **-79.28%** |
| Old | 1.060 | 1.056 | 1.056 | **1.056s**|  |  |
| New | 0.212 | 0.215 | 0.215 | **0.215s** | **-0.841s** | **-79.62%** |
| Old | 0.206 | 0.205 | 0.206 | **0.205s**|  |  |
| New | 0.044 | 0.045 | 0.045 | **0.046s** | **-0.159s** | **-77.77%** |
| Old | 0.856 | 0.855 | 0.856 | **0.855s**|  |  |
| New | 0.173 | 0.172 | 0.180 | **0.177s** | **-0.678s** | **-79.32%** |
| Old | 1.430 | 1.431 | 1.430 | **1.431s**|  |  |
| New | 0.291 | 0.286 | 0.291 | **0.292s** | **-1.139s** | **-79.6%** |
| Old | 1.785 | 1.780 | 1.780 | **1.780s**|  |  |
| New | 0.360 | 0.355 | 0.363 | **0.362s** | **-1.418s** | **-79.66%** |
| Old | 1.107 | 1.105 | 1.106 | **1.105s**|  |  |
| New | 0.225 | 0.222 | 0.223 | **0.225s** | **-0.880s** | **-79.65%** |
| Old | 1.481 | 1.480 | 1.481 | **1.479s**|  |  |
| New | 0.298 | 0.296 | 0.307 | **0.300s** | **-1.179s** | **-79.7%** |
| Old | 1.681 | 1.684 | 1.680 | **1.682s**|  |  |
| New | 0.335 | 0.335 | 0.336 | **0.338s** | **-1.344s** | **-79.92%** |
| Old | 1.964 | 1.957 | 1.957 | **1.957s**|  |  |
| New | 0.390 | 0.391 | 0.396 | **0.395s** | **-1.561s** | **-79.79%** |
| Old | 0.656 | 0.655 | 0.655 | **0.655s**|  |  |
| New | 0.133 | 0.133 | 0.141 | **0.136s** | **-0.519s** | **-79.26%** |
| Old | 0.131 | 0.130 | 0.131 | **0.130s**|  |  |
| New | 0.030 | 0.030 | 0.031 | **0.030s** | **-0.100s** | **-76.73%** |
| Old | 7.464 | 7.467 | 7.457 | **7.466s**|  |  |
| New | 1.476 | 1.476 | 1.496 | **1.490s** | **-5.976s** | **-80.05%** |
| Old | 0.258 | 0.255 | 0.255 | **0.256s**|  |  |
| New | 0.055 | 0.055 | 0.054 | **0.055s** | **-0.201s** | **-78.46%** |
| Old | 0.182 | 0.181 | 0.181 | **0.181s**|  |  |
| New | 0.040 | 0.039 | 0.040 | **0.040s** | **-0.142s** | **-78.13%** |
| Old | 1.882 | 1.881 | 1.881 | **1.885s**|  |  |
| New | 0.375 | 0.378 | 0.385 | **0.380s** | **-1.505s** | **-79.85%** |
| Old | 3.112 | 3.107 | 3.104 | **3.112s**|  |  |
| New | 0.617 | 0.626 | 0.630 | **0.623s** | **-2.488s** | **-79.97%** |
| Old | 1.309 | 1.306 | 1.306 | **1.308s**|  |  |
| New | 0.266 | 0.264 | 0.270 | **0.266s** | **-1.042s** | **-79.65%** |
| Old | 1.580 | 1.582 | 1.582 | **1.582s**|  |  |
| New | 0.321 | 0.316 | 0.318 | **0.319s** | **-1.263s** | **-79.83%** |
| Old | 0.481 | 0.481 | 0.481 | **0.481s**|  |  |
| New | 0.100 | 0.099 | 0.100 | **0.100s** | **-0.381s** | **-79.23%** |
| Old | 1.611 | 1.605 | 1.606 | **1.607s**|  |  |
| New | 0.326 | 0.321 | 0.326 | **0.324s** | **-1.283s** | **-79.83%** |
| Old | 0.757 | 0.756 | 0.755 | **0.757s**|  |  |
| New | 0.158 | 0.154 | 0.158 | **0.156s** | **-0.601s** | **-79.38%** |
| Old | 0.531 | 0.531 | 0.531 | **0.532s**|  |  |
| New | 0.111 | 0.108 | 0.113 | **0.110s** | **-0.421s** | **-79.25%** |
| Old | 0.706 | 0.705 | 0.706 | **0.706s**|  |  |
| New | 0.146 | 0.144 | 0.145 | **0.145s** | **-0.561s** | **-79.47%** |
| Old | 3.137 | 3.131 | 3.131 | **3.136s**|  |  |
| New | 0.623 | 0.624 | 0.633 | **0.630s** | **-2.506s** | **-79.92%** |
| Old | 2.082 | 2.081 | 2.081 | **2.082s**|  |  |
| New | 0.415 | 0.415 | 0.423 | **0.420s** | **-1.663s** | **-79.85%** |
| Old | 0.781 | 0.781 | 0.781 | **0.782s**|  |  |
| New | 0.158 | 0.158 | 0.161 | **0.160s** | **-0.622s** | **-79.51%** |
| Old | 0.206 | 0.206 | 0.205 | **0.206s**|  |  |
| New | 0.045 | 0.045 | 0.046 | **0.045s** | **-0.160s** | **-77.93%** |
| Old | 5.762 | 5.758 | 5.759 | **5.761s**|  |  |
| New | 1.139 | 1.139 | 1.197 | **1.159s** | **-4.601s** | **-79.87%** |

That is using 8X diffculty. Using 1X drops to about 0.05 seconds per PoW. _The "Old" is not the same as #1 , because there I use the GPU+CPU, and now it's just GPU. Switching from CPU+GPU to GPU changes the initial nonce, which changes the time per work._

---------

It's almost 80% faster now. 🥇